oscillillator offby1 bug, iir initialize last_xh, use make_unique for buffers in FFT and IIR

+++ Application/SamplingTypes.h
@@ -1,1 +1,1 @@
-typedef double sample_t ;
+typedef float sample_t ;
